+++ src/WettergeraetLib/FetchUrl.cpp
@@ -129,6 +129,11 @@ std::string FetchUrl::fetch(std::string url, std::string icaoCode, std::string d
 		std::regex("[^a-zA-z0-9]"),
 		""
 	);
+	date = std::regex_replace(
+		date,
+		std::regex(":"),
+		std::string("%3A")
+	);
 	if (lowercase) {
 		transform(icaoCode.begin(), icaoCode.end(), icaoCode.begin(), ::tolower);
 	}